Write distance range for depthmap encoding to file
The distance range set in the Acquisition component is stored in
"additional_information.txt" so it can be recovered later when
processing the depthmaps. Otherwise decoding the depth values leads
to incorrect results.

Conceptually, the distance range should be stored in "images.txt",
but that would interfere with the file's COLMAP format.

Although it appears that the distance range can only be set globally,
it is stored on the CameraModel level, and so an individual distance
range pair is written for each camera.